Split out CalculateSectionalStrain
